+++ tests/test_models/test_inference/test_predict.py
@@ -31,6 +31,8 @@ from etna.models import StatsForecastAutoCESModel
 from etna.models import StatsForecastAutoETSModel
 from etna.models import StatsForecastAutoThetaModel
 from etna.models import TBATSModel
+from etna.models.nn import ChronosBoltModel
+from etna.models.nn import ChronosModel
 from etna.models.nn import DeepARModel
 from etna.models.nn import DeepARNativeModel
 from etna.models.nn import DeepStateModel
@@ -191,6 +193,8 @@ class TestPredictInSampleFull:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_in_sample_full_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -320,6 +324,8 @@ class TestPredictInSampleSuffix:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_in_sample_suffix_datetime_timestamp_failed_not_implemented_predict(
@@ -464,6 +470,8 @@ class TestPredictInSampleSuffix:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_in_sample_suffix_int_timestamp_failed_not_implemented_predict(
@@ -604,6 +612,8 @@ class TestPredictOutSample:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_out_sample_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -761,6 +771,8 @@ class TestPredictOutSamplePrefix:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_out_sample_prefix_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -932,6 +944,8 @@ class TestPredictOutSampleSuffix:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_out_sample_suffix_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1108,6 +1122,8 @@ class TestPredictMixedInOutSample:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_mixed_in_out_sample_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1272,6 +1288,8 @@ class TestPredictSubsetSegments:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_subset_segments_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1401,6 +1419,8 @@ class TestPredictNewSegments:
                 "example_tsds",
             ),
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
+            (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
+            (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
         ],
     )
     def test_predict_new_segments_failed_not_implemented_predict(self, model, transforms, dataset_name, request):